refactor: fetch data to action directory
