+++ src/components/lesseeHome-page/propertyScroll.tsx
@@ -1,4 +1,4 @@
-import { fetchProperty } from '@/actions/fetchProperty';
+import { fetchProperty } from '@/actions/lesseeView/fetchProperty';
 import type { components } from '@/types/api';
 import PaginationControl from '@/components/lesseeHome-page/paginationControl';
 import { redirect } from 'next/navigation';
@@ -17,10 +17,11 @@ export async function PropertyScroll({
   // console.log(page)
   const data = await fetchProperty(page);
 
-  const propertyData = !data ? JSON.parse(mockData) : data;
+  const propertyData = data ? data : JSON.parse(mockData);
   if (propertyData.length == 0 || page < 1 || Number.isNaN(page)) {
     redirect(lesseePagePath);
   }
+  console.log(propertyData);
   return (
     <div className="shadow-md border border-gray-100 pt-14 pb-12">
       <p className="block text-center text-lg w-[150px] mb-8 ml-2 rounded-2xl border border-gray-400">
